+++ bot/category.py
@@ -173,6 +173,18 @@ class PollsCategory(Category):
     def check_permissions(self, message: discord.Message) -> bool:
         return True
 
+class MinecraftCategory(Category):
+    """A command category instance."""
+
+    name = "minecraft"
+    prefix = "minecraft"
+    commands: List[Command] = []
+
+    def check_permissions(self, message: discord.Message) -> bool:
+        return True
+
+
+
 
 if __name__ == "__main__":
     print(
